moved ChunkMapObjectArrayCopy into its own header
